% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobiEquity.R
\docType{data}
\name{sobi_hubs}
\alias{sobi_hubs}
\title{Hamilton Bike Share Stations}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 133 rows and 14 columns.
}
\source{
Open Hamilton (https://open.hamilton.ca/)
}
\usage{
data(sobi_hubs)
}
\description{
Location of the Hamilton Bike Share stations of Hamilton, in Ontario, Canada.
}
\keyword{datasets}
\keyword{transportation}
