% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobiEquity.R
\docType{package}
\name{sobiEquity}
\alias{sobiEquity}
\title{sobiEquity}
\description{
sobiEquity is a companion package for the paper "Examining horizontal and vertical equity to a public bicycle share program using a balanced floating catchment area accessibility approach".
  The package includes the data used in the analysis.
}
\author{
Elise Desjardins, \email{desjae@mcmaster.ca}

Antonio Paez, \email{paezha@mcmaster.ca}
}
